preserve encoding over temporary storage
little bit of refactoring over temporary stroage and text factory service to preserve encoding round trip
