+++ src/EditorFeatures/Text/Extensions.SnapshotSourceText.cs
@@ -28,8 +28,8 @@ namespace Microsoft.CodeAnalysis.Text
             /// </summary>
             internal class ClosedSnapshotSourceText : SnapshotSourceText
             {
-                public ClosedSnapshotSourceText(ITextSnapshot roslynSnapshot, Encoding encoding)
-                    : base(roslynSnapshot, encoding, containerOpt: null)
+                public ClosedSnapshotSourceText(ITextSnapshot roslynSnapshot, Encoding encodingOpt)
+                    : base(roslynSnapshot, encodingOpt, containerOpt: null)
                 {
                 }
             }
@@ -40,29 +40,27 @@ namespace Microsoft.CodeAnalysis.Text
             /// The ITextSnapshot backing the SourceText instance
             /// </summary>
             protected readonly ITextSnapshot RoslynSnapshot;
-            private readonly Encoding _encoding;
+            private readonly Encoding _encodingOpt;
             private readonly TextBufferContainer _containerOpt;
             private readonly int _reiteratedVersion;
             private LineInfo _lineInfo;
 
-            private SnapshotSourceText(ITextSnapshot editorSnapshot, Encoding encoding)
+            private SnapshotSourceText(ITextSnapshot editorSnapshot, Encoding encodingOpt)
             {
                 Contract.ThrowIfNull(editorSnapshot);
-                Contract.ThrowIfNull(encoding);
 
                 this.RoslynSnapshot = TextBufferMapper.ToRoslyn(editorSnapshot);
                 _containerOpt = TextBufferContainer.From(editorSnapshot.TextBuffer);
                 _reiteratedVersion = editorSnapshot.Version.ReiteratedVersionNumber;
-                _encoding = encoding;
+                _encodingOpt = encodingOpt;
             }
 
-            public SnapshotSourceText(ITextSnapshot roslynSnapshot, Encoding encoding, TextBufferContainer containerOpt)
+            public SnapshotSourceText(ITextSnapshot roslynSnapshot, Encoding encodingOpt, TextBufferContainer containerOpt)
             {
                 Contract.ThrowIfNull(roslynSnapshot);
-                Contract.ThrowIfNull(encoding);
 
                 this.RoslynSnapshot = roslynSnapshot;
-                _encoding = encoding;
+                _encodingOpt = encodingOpt;
                 _containerOpt = containerOpt;
             }
 
@@ -101,7 +99,7 @@ namespace Microsoft.CodeAnalysis.Text
 
             public override Encoding Encoding
             {
-                get { return _encoding; }
+                get { return _encodingOpt; }
             }
 
             public ITextSnapshot EditorSnapshot
